feat(W44-213): wire RuntimeTuning into 11 production encoder consumer sites
Closes the W44-212 SCAFFOLDING_NOTE caveat: until this commit, installing
a non-default RuntimeTuning via tuning::runtime::install was a no-op at
the encoder, so the W44-212 sweep runner captured params_blob in Parquet
but encoded_bytes was independent of tuning-axis variance — useless for
MLP training.

## Mechanism

1. New `runtime_or_default!` macro in `src/tuning.rs`:
   - Feature `tuning-override` OFF (production default): expands to the
     const reference (zero overhead, compiler inlines).
   - Feature ON: expands to `tuning::runtime::get_or_default(default, accessor)`
     which short-circuits to the default when no override is installed
     (single atomic-OnceLock load + branch).

2. New `runtime::get_or_default` fn alongside the existing `get`
   helper — supplies an explicit default so the fast-path (no install)
   never needs to construct `RuntimeTuning::default()`.

3. Wired all 6 RuntimeTuning fields at 11 production consumer sites:

   | field | sites |
   |---|---|
   | smart_zenjxl_photo_mask_p25_min | encoder.rs:929, 2832, 2880/2883, 4417, 4543/4546 (5 sites) |
   | screenshot_median_threshold | encoder.rs:929, 2755, 3920, 4277, 5055 (5 sites) |
   | buttloop_default_screenshot_qf_seed_scale | butteraugli_loop.rs:1361 (1 site) |
   | buttloop_qf_seed_scale_min_distance | butteraugli_loop.rs:758, 1348 (2 sites) |
   | adaptive_quant_screenshot_qf_seed_scale_e5_e6 | butteraugli_loop.rs:793 (1 site) |
   | adaptive_quant_screenshot_qf_seed_scale_e7 | butteraugli_loop.rs:790 (1 site) |

## Verification

- (a) Build PASS (default + `--features tuning-override`)
- (b) Library tests: 1405 default → 1407 with feature
- (c) Hash-locks 36/36 BYTE-IDENTICAL at default AND with feature
- (d) Libjxl byte-locks 4/4 PASS at default AND with feature
- (e) divergence_table_drift 7/7 PASS
- (f) NEW wiring-proof test
  (`jxl-encoder/tests/w44_213_runtime_tuning_wiring.rs`): installs a
  non-default RuntimeTuning that doubles
  `buttloop_default_screenshot_qf_seed_scale` (4.0 → 8.0); encoded
  bytes for `gb82-sc/terminal.png` (clipped to 512×512) at e8 d=4
  change from 8904 → 11651 (+30.85%), proving the consumer site
  actually consults the runtime override.
- (g) `docs/TUNING_RELATIONS.md` Section 0: added W44-213 wired-fields
  table with per-field consumer-site map; removed W44-212 scaffolding
  caveat.
- (h) `zenjxl-tuning-runner/src/lib.rs` SCAFFOLDING_NOTE updated to
  document the closed gap.
- (i) `docs/LIBJXL_DIVERGENCES.md` header notes the W44-213 wiring
  (not a divergence row — no production-default behaviour change).

## Production-binary safety

- `runtime_or_default!` with feature OFF compiles down to the
  immediate const value at every call site. Zero overhead.
- `RuntimeTuning::default()` matches every source-of-truth const
  exactly, enforced by
  `tuning::runtime::tests::default_matches_production_consts`.
- Hash-locks 36/36 byte-identical at the `tuning-override` feature
  default (i.e. no override installed).

## Follow-on candidates (NOT this chunk)

- Extend RuntimeTuning to additional tunables (W44_91_M3_COLOURFULNESS_MIN,
  W44_124_DCT32_KEEP_*, LIBJXL_INIT_MUL, DEFAULT_CUR_POW_LOW, EPF
  sharpness thresholds) — same wiring pattern.
- W44-214+: actual MLP training on the params_blob × encoded_bytes
  joins now that the sweep produces meaningful variance.

+++ zenjxl-tuning-runner/src/parquet_writer.rs
@@ -26,8 +26,8 @@ use std::path::Path;
 use std::sync::Arc;
 
 use arrow_array::{
-    ArrayRef, BinaryArray, Float32Array, Float64Array, RecordBatch, StringArray, UInt32Array,
-    UInt64Array, UInt8Array,
+    ArrayRef, BinaryArray, Float32Array, Float64Array, RecordBatch, StringArray, UInt8Array,
+    UInt32Array, UInt64Array,
 };
 // Needed in tests; re-export at module scope so `cargo test` doesn't
 // warn about an unused import behind `#[cfg(test)]`.
@@ -52,11 +52,14 @@ pub fn write_single_row_parquet(row: &SweepCellRow, path: &Path) -> Result<u64, 
     let batch = RecordBatch::try_new(schema.clone(), arrays)
         .map_err(|e| format!("RecordBatch::try_new: {e}"))?;
 
-    let file = std::fs::File::create(path)
-        .map_err(|e| format!("create {}: {e}", path.display()))?;
+    let file =
+        std::fs::File::create(path).map_err(|e| format!("create {}: {e}", path.display()))?;
     let props = WriterProperties::builder()
         .set_compression(Compression::ZSTD(ZstdLevel::try_new(3).unwrap()))
-        .set_created_by(format!("zenjxl-tuning-runner v{}", env!("CARGO_PKG_VERSION")))
+        .set_created_by(format!(
+            "zenjxl-tuning-runner v{}",
+            env!("CARGO_PKG_VERSION")
+        ))
         .build();
     let mut writer = ArrowWriter::try_new(file, schema, Some(props))
         .map_err(|e| format!("ArrowWriter::try_new: {e}"))?;
